+++ fiberplane-hooks/src/useHandler.ts
@@ -2,6 +2,7 @@ import { useCallback, useRef } from "react";
 
 const noDeps: Array<void> = [];
 
+// biome-ignore lint/suspicious/noExplicitAny: not aware of a better solution in TS
 export function useHandler<Handler extends (...args: Array<any>) => any>(
   handler: Handler,
 ): Handler {